refactor(db): extract read cursor repository
